Move response parsing to module in hoverkite-protocol.
For now it still requires std.
