+++ hovercontrol/Cargo.toml
@@ -13,6 +13,3 @@ pretty_env_logger = "0.4.0"
 serialport = "4.0.0"
 stable-eyre = "0.2.2"
 hoverkite-protocol = { path = "../hoverkite-protocol" }
-
-[dev-dependencies]
-test-case = "1.1.0"
